feat(provider): config-first for ollama and groq

+++ siumai-provider-groq/src/providers/groq/config.rs
@@ -47,7 +47,7 @@ impl GroqConfig {
     }
 
     /// Set the temperature
-    pub fn with_temperature(mut self, temperature: f32) -> Self {
+    pub fn with_temperature(mut self, temperature: f64) -> Self {
         self.common_params.temperature = Some(temperature);
         self
     }
@@ -59,7 +59,7 @@ impl GroqConfig {
     }
 
     /// Set the top_p parameter
-    pub fn with_top_p(mut self, top_p: f32) -> Self {
+    pub fn with_top_p(mut self, top_p: f64) -> Self {
         self.common_params.top_p = Some(top_p);
         self
     }
